Improve and document CI/CD auth check (#657)
Based on some feedback from others, it would be helpful to have helper
scripts to check the status of various stages of the infra template
setup.

This change improves the ci/cd auth check to be easier to use by only
needing the account name rather than the role_arn, and added steps to
the set-up-aws-account docs. This way devs can check each step as they
progress through the infra template setup.

---------

Co-authored-by: kai [they] <coilysiren@gmail.com>